[forwarder] handle errors for each endpoint
Previously everything was dismissed except for the last result.
This commit duplicates payload for each endpoint, so that the transaction
manager handles them properly.

Another solution would have been to make the transaction aware of the errors
for each of the endpoints, but the Transaction Manager and the Transaction were
not created to do this, and thus is/would be a huge pain to update them.
(because of the async nature of the flush)